linux64-requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
asn1crypto=0.24.0=py36_0
blas=1.0=mkl
boto=2.49.0=py36_0
boto3=1.9.35=py36_0
botocore=1.12.35=py36_0
bottleneck=1.2.1=py36h035aef0_1
ca-certificates=2019.1.23=0
certifi=2018.11.29=py36_0
cffi=1.11.5=py36he75722e_1
chardet=3.0.4=py36_1
cryptography=2.5=py36h1ba5d50_0
cryptography-vectors=2.3.1=py36_1000
cuda80=1.0=h205658b_0
cycler=0.10.0=py36_0
cymem=2.0.2=py36hfd86e86_0
cytoolz=0.9.0.1=py36h14c3975_1
dataclasses=0.6=py_0
dbus=1.13.2=h714fa37_1
dill=0.2.8.2=py36_0
docutils=0.14=py36_0
expat=2.2.6=he6710b0_0
fastai=1.0.36=py_1
fastprogress=0.1.18=py_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
gensim=3.4.0=py36h14c3975_0
glib=2.56.2=hd408876_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
icu=58.2=h211956c_0
idna=2.7=py36_0
intel-openmp=2019.1=144
jmespath=0.9.3=py36_0
jpeg=9b=habf39ab_1
kiwisolver=1.0.1=py36hf484d3e_0
libedit=3.1.20170329=h6b74fdf_2
libffi=3.2.1=h4deb6c0_3
libgcc=7.2.0=h69d50b8_2
libgcc-ng=8.2.0=hdf63c60_1
libgfortran-ng=7.3.0=hdf63c60_0
libjpeg-turbo=2.0.1=h470a237_0
libpng=1.6.35=hbc83047_0
libstdcxx-ng=8.2.0=hdf63c60_1
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.8=h26e45fe_1
matplotlib=3.0.2=py36h5429711_0
mkl=2019.1=144
mkl_fft=1.0.6=py36h7dd41cf_0
mkl_random=1.0.1=py36h4414c95_1
msgpack-numpy=0.4.3.2=py36_0
msgpack-python=0.5.6=py36h6bb024c_1
murmurhash=1.0.1=py36he6710b0_0
ncurses=6.1=he6710b0_1
ninja=1.8.2=py36h6bb024c_1
numexpr=2.6.8=py36hd89afb7_0
numpy=1.15.4=py36h1d66e8a_0
numpy-base=1.15.4=py36h81de0dd_0
olefile=0.46=py36_0
openssl=1.1.1=h7b6447c_0
pandas=0.23.4=py36h04863e7_0
pcre=8.42=h439df22_0
pillow=5.3.0=py36h34e0f95_0
pip=18.1=py36_0
plac=0.9.6=py36_0
preshed=2.0.1=py36he6710b0_0
pycparser=2.19=py36_0
pyopenssl=18.0.0=py36_0
pyparsing=2.3.0=py36_0
pyqt=5.9.2=py36h22d08a2_1
pysocks=1.6.8=py36_0
python=3.6.7=h0371630_0
python-dateutil=2.7.5=py36_0
pytorch=1.0.0=py3.6_cuda8.0.61_cudnn7.1.2_1
pytz=2018.7=py36_0
pyyaml=3.13=py36h14c3975_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
regex=2018.08.29=py36h7b6447c_0
requests=2.20.1=py36_0
s3transfer=0.1.13=py36_0
scikit-learn=0.20.1=py36h4989274_0
scipy=1.1.0=py36hfa4b5c9_1
setuptools=40.6.2=py36_0
sip=4.19.13=py36he6710b0_0
six=1.11.0=py36_1
smart_open=1.7.1=py36_0
spacy=2.0.16=py36h962f231_0
sqlite=3.25.3=h7b6447c_0
thinc=6.12.0=py36h4989274_0
tk=8.6.8=hbc83047_0
toolz=0.9.0=py36_0
torchvision=0.2.1=py_2
tornado=5.1.1=py36h7b6447c_0
tqdm=4.28.1=py36h28b3542_0
typing=3.6.4=py36_0
ujson=1.35=py36h14c3975_0
urllib3=1.23=py36_0
wheel=0.32.3=py36_0
wrapt=1.10.11=py36h14c3975_2
xz=5.2.4=h14c3975_4
yaml=0.1.7=h96e3832_1
zlib=1.2.11=h7b6447c_3
